[CCUBE-1436][SR] Amend the importing, extract out the interfaces in translator, removed unused imports, amended the prefill translation.
