+++ src/translator/index.ts
@@ -1,2 +1,4 @@
 export * from "./translator";
 export * from "./helper";
+export * from "./types";
+export * from "./text-based-field";
